+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Since last release
 
 **Added:**
 
+* Added progress bar to the simulation loop (#1912)
 * Added a warning for when a facility trades with itself (#1895)
 * Added a new datatype to the backend for tariff region (#1922)
 * Added Functions to access region and institution from facilities and institutions (#1914)